requirements.txt
slack-sdk==3.23.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
click==8.1.7
tqdm==4.66.1
pytz==2023.3
